test: wait after adding disk
We were adding a disk and using `dd if=/dev/sda of=/dev/sda` to simulate
disk activity for testing the metrics backends.  We never actually
waited to make sure the disk came online.

This is an existing issue that was revealed by a timing change: the new
way we run scripts via Machine.execute() is apparently a bit faster.